+++ include/SI/detail/eps_equal.h
@@ -1,0 +1,28 @@
+/**
+ * This file is part of "SI" version 2.1.0
+ * A header only c++ library that provides type safety and user defined literals
+ * for handling pyhsical values defined in the International System of
+ * Units
+ *
+ * https://github.com/bernedom/SI
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ **/
+#pragma once
+
+#include <limits>
+#include <numeric>
+#include <type_traits>
+
+namespace SI::detail {
+
+template <typename T, std::enable_if_t<std::is_floating_point_v<T>> * = nullptr>
+constexpr bool eps_equals(const T &lhs, const T &rhs) {
+
+  return (lhs - rhs) < std::numeric_limits<T>::epsilon() &&
+         (lhs - rhs) > -std::numeric_limits<T>::epsilon();
+  // return std::abs(lhs - rhs) < std::numeric_limits<T>::epsilon();
+}
+
+} // namespace SI::detail