+++ include/SI/detail/unit.h
@@ -1,5 +1,5 @@
 /**
- * This file is part of "SI" version 2.0.5
+ * This file is part of "SI" version 2.1.0
  * A header only c++ library that provides type safety and user defined literals
  * for handling pyhsical values defined in the International System of
  * Units
@@ -18,16 +18,14 @@
 #endif
 
 #include "detail.h"
+#include "eps_equal.h"
+#include "unit_cast.h"
 
 #include <ratio>
 #include <type_traits>
 
 /// Namespace containing all SI units
 namespace SI::detail {
-
-// forward declarations
-template <typename _target_type, typename _rhs_T>
-constexpr auto unit_cast(const _rhs_T &rhs);
 
 template <typename _unit_lhs, typename _unit_rhs> struct unit_with_common_ratio;
 
@@ -168,8 +166,8 @@ struct unit_t {
       return unit_cast<gcd_unit>(rhs).value() ==
              unit_cast<gcd_unit>(*this).value();
     } else {
-      return detail::epsEqual(unit_cast<gcd_unit>(rhs).value(),
-                              unit_cast<gcd_unit>(*this).value());
+      return detail::eps_equals(unit_cast<gcd_unit>(rhs).value(),
+                                unit_cast<gcd_unit>(*this).value());
     }
   }
 
@@ -497,93 +495,6 @@ operator/(const _type &lhs,
              std::ratio<1>>{lhs /
                             (rhs.value() * (static_cast<_type>(_ratio::num) /
                                             static_cast<_type>(_ratio::den)))});
-}
-
-/// helper template to check if a type is a unit_t (false for all other types)
-template <typename _unit> struct is_unit_t : std::false_type {};
-
-/// template specialisation to check if a type is a unit_t (true if unit_t)
-template <char _symbol, typename _exponent, typename _ratio, typename _type>
-struct is_unit_t<const unit_t<_symbol, _exponent, _type, _ratio>>
-    : std::true_type {};
-
-/// non-const specialisation of check above
-template <char _symbol, typename _exponent, typename _ratio, typename _type>
-struct is_unit_t<unit_t<_symbol, _exponent, _type, _ratio>> : std::true_type {};
-
-template <typename _type>
-inline constexpr bool is_unit_t_v = is_unit_t<_type>::value;
-
-/// function to cast between two units of the same type
-template <typename _target_type, typename _rhs_T>
-constexpr auto unit_cast(const _rhs_T &rhs) {
-  // using static assert instead of std::enable if in order to be able to
-  // forward declare this function easier
-  static_assert(
-      is_unit_t_v<_rhs_T> ||
-          std::is_base_of<
-              unit_t<_rhs_T::symbol::value, typename _rhs_T::exponent,
-                     typename _rhs_T::internal_type, typename _rhs_T::ratio>,
-              _rhs_T>::value,
-      "is of type unit_t or a derived class");
-  using conversion_ratio =
-      std::ratio_divide<typename _rhs_T::ratio, typename _target_type::ratio>;
-
-  return _target_type(
-      ((rhs.value() * conversion_ratio::num) / conversion_ratio::den));
-}
-
-template <typename _unit_lhs, typename _unit_rhs>
-struct unit_with_common_ratio {
-  static_assert(is_unit_t_v<_unit_lhs>, "only supported for SI::unit_t");
-  static_assert(is_unit_t_v<_unit_rhs>, "only supported for SI::unit_t");
-  static_assert(std::is_convertible<typename _unit_lhs::internal_type,
-                                    typename _unit_rhs::internal_type>::value);
-  static_assert(_unit_lhs::symbol::value == _unit_rhs::symbol::value);
-  using type =
-      unit_t<_unit_lhs::symbol::value, typename _unit_lhs::exponent,
-             typename _unit_lhs::internal_type,
-             typename detail::ratio_gcd<typename _unit_lhs::ratio,
-                                        typename _unit_rhs::ratio>::ratio>;
-};
-
-/// divide a value of a certain unit with another value of a possibly
-/// different type resulting in a new type, the resulting exponent is
-/// specified by resulting unit using a variadic template to simplify usage of
-/// implementation the internal type of the result is the internal type of lhs
-template <template <typename...> typename _resulting_unit, typename _unit_lhs,
-          typename _unit_rhs>
-constexpr auto cross_unit_divide(const _unit_lhs &lhs, const _unit_rhs &rhs) {
-  // do not use for the same unit as this should result in decreasing the
-  // exponent
-  static_assert(!std::is_same<_unit_lhs, _unit_rhs>::value);
-  static_assert(is_unit_t_v<_unit_lhs>, "lhs parameter is a unit_t");
-  static_assert(is_unit_t_v<_unit_rhs>, "rhs parameter is a unit_t");
-
-  using resulting_ratio = typename std::ratio_divide<typename _unit_lhs::ratio,
-                                                     typename _unit_rhs::ratio>;
-  return _resulting_unit<typename _unit_lhs::internal_type, resulting_ratio>(
-      lhs.value() / rhs.value());
-}
-/// multiply a value of a unit witn another value of a possibly different
-/// value resulting in a value of a new type with exponent 1 the internal type
-/// of the result is the internal type of lhs
-/// @todo add function that works with variable exponent units and remove
-/// special typedefs for time
-
-template <template <typename...> typename _resulting_unit, typename _unit_lhs,
-          typename _unit_rhs>
-constexpr auto cross_unit_multiply(const _unit_lhs &lhs, const _unit_rhs &rhs) {
-  // do not use for the same unit as this should result in increasing the
-  // exponent
-  static_assert(!std::is_same<_unit_lhs, _unit_rhs>::value);
-  static_assert(is_unit_t_v<_unit_lhs>, "lhs parameter is a unit_t");
-  static_assert(is_unit_t_v<_unit_rhs>, "rhs parameter is a unit_t");
-  using resulting_ratio =
-      typename std::ratio_multiply<typename _unit_lhs::ratio,
-                                   typename _unit_rhs::ratio>;
-  return _resulting_unit<typename _unit_lhs::internal_type, resulting_ratio>(
-      lhs.value() * rhs.value());
 }
 
 } // namespace SI::detail